Move subbasins script into scripts directory
